Consolidate listify/flatten in one place
