+++ src/gt4py/cartesian/gtc/dace/expansion/daceir_builder.py
@@ -17,7 +17,8 @@ import dace
 import dace.data
 
 from gt4py import eve
-from gt4py.cartesian.gtc import common, oir, utils as gtc_utils
+from gt4py.cartesian import utils as gt_utils
+from gt4py.cartesian.gtc import common, oir
 from gt4py.cartesian.gtc.dace import daceir as dcir
 from gt4py.cartesian.gtc.dace.expansion.utils import remove_horizontal_region
 from gt4py.cartesian.gtc.dace.expansion_specification import Loop, Map, Sections, Stages
@@ -608,7 +609,7 @@ class DaCeIRBuilder(eve.NodeTranslator):
             **kwargs,
         )
         targets.clear()
-        return dcir.Condition(condition=tasklet, true_states=gtc_utils.listify(code_block))
+        return dcir.Condition(condition=tasklet, true_states=gt_utils.listify(code_block))
 
     def visit_While(
         self,
