+++ src/gt4py/cartesian/gtc/numpy/oir_to_npir.py
@@ -9,7 +9,8 @@
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 from gt4py import eve
-from gt4py.cartesian.gtc import common, oir, utils
+from gt4py.cartesian import utils
+from gt4py.cartesian.gtc import common, oir
 from gt4py.cartesian.gtc.definitions import Extent
 from gt4py.cartesian.gtc.numpy import npir
 from gt4py.cartesian.gtc.passes.horizontal_masks import compute_relative_mask
@@ -132,7 +133,7 @@ class OirToNpir(eve.NodeTranslator, eve.VisitorWithSymbolTableTrait):
         if mask:
             mask_expr = npir.VectorLogic(op=common.LogicalOperator.AND, left=mask, right=mask_expr)
 
-        return utils.flatten_list(self.visit(node.body, mask=mask_expr, **kwargs))
+        return utils.flatten(self.visit(node.body, mask=mask_expr, **kwargs))
 
     def visit_AssignStmt(
         self,
@@ -161,7 +162,7 @@ class OirToNpir(eve.NodeTranslator, eve.VisitorWithSymbolTableTrait):
             cond_expr = npir.VectorLogic(op=common.LogicalOperator.AND, left=mask, right=cond_expr)
 
         return npir.While(
-            cond=cond_expr, body=utils.flatten_list(self.visit(node.body, mask=cond_expr, **kwargs))
+            cond=cond_expr, body=utils.flatten(self.visit(node.body, mask=cond_expr, **kwargs))
         )
 
     def visit_HorizontalRestriction(
@@ -173,7 +174,7 @@ class OirToNpir(eve.NodeTranslator, eve.VisitorWithSymbolTableTrait):
 
         horizontal_mask = npir.HorizontalMask(i=mask[0], j=mask[1])
 
-        return utils.flatten_list(self.visit(node.body, horizontal_mask=horizontal_mask, **kwargs))
+        return utils.flatten(self.visit(node.body, horizontal_mask=horizontal_mask, **kwargs))
 
     # --- Control Flow ---
     def visit_HorizontalExecution(
@@ -188,7 +189,7 @@ class OirToNpir(eve.NodeTranslator, eve.VisitorWithSymbolTableTrait):
         else:
             extent = Extent.zeros(ndims=2)
 
-        stmts = utils.flatten_list(self.visit(node.body, extent=extent, **kwargs))
+        stmts = utils.flatten(self.visit(node.body, extent=extent, **kwargs))
         return npir.HorizontalBlock(
             body=stmts, extent=extent, declarations=self.visit(node.declarations, **kwargs)
         )
@@ -222,7 +223,7 @@ class OirToNpir(eve.NodeTranslator, eve.VisitorWithSymbolTableTrait):
             self.visit(decl, field_extents=field_extents, **kwargs) for decl in node.declarations
         ]
 
-        vertical_passes = utils.flatten_list(
+        vertical_passes = utils.flatten(
             self.visit(node.vertical_loops, block_extents=block_extents, **kwargs)
         )
