+++ src/gt4py/cartesian/utils/__init__.py
@@ -24,6 +24,7 @@ from .base import (
     is_iterable_of,
     is_mapping_of,
     jsonify,
+    listify,
     make_dir,
     make_local_dir,
     make_module_from_file,
@@ -59,6 +60,7 @@ __all__ = [  # noqa: RUF022 `__all__` is not sorted
     "is_iterable_of",
     "is_mapping_of",
     "jsonify",
+    "listify",
     "make_dir",
     "make_local_dir",
     "make_module_from_file",
